pyproject.toml
[tool.poetry]
name = "stpyapp-template"
version = "0.1.0"
description = "streamlit project template for quick start"
authors = ["Shogo Ogami <sg.tao.so@gmail.com>"]
license = "Apache-2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
streamlit = "^1.35.0"
altair = "^5.3.0"
pandas = "^2.2.2"
pillow = "^10.3.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
taskipy = "^1.13.0"
flake8 = "^7.1.0"
black = "^24.4.2"
pytest-cov = "^5.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.taskipy.tasks]
start = "streamlit run src/main.py"
test = "pytest tests"
test-cov = "pytest tests --cov --cov-branch -svx"
test-repo = "pytest tests --cov --cov-report=html"
format = "black --line-length 79 src"
lint = "flake8 src"
